+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## Recent Changes
+
+- Add `encoding` option to `zosmf` profile type.
+
 ## `6.14.0`
 
 - Add encoding / code page support for data set upload and download operations in library and CLI.